requirements.txt
dlib-bin>=19.16.0
imutils>=0.5.2
numpy>=1.15.4
opencv-python>=4.0.0.21
pygame>=2.0.0
scipy>=1.2.0

# Note: playsound has compatibility issues with Python 3.13
# pygame is used as a fallback for audio playback
# If you need playsound, try: pip install playsound==1.3.0 (may require Python < 3.13)